Add initTestHelpers
