+++ src/index.ts
@@ -1,1 +1,2 @@
 export { default as getPage } from './getPage';
+export { initTestHelpers } from './testHelpers';
